requirements.txt
Flask==2.0.1
Werkzeug==2.0.3
Flask-SQLAlchemy==2.5.1
Flask-Cors==3.0.10
SQLAlchemy==1.4.23
pg8000==1.30.3
gunicorn==20.1.0
python-dotenv==0.19.0
requests==2.26.0
PyJWT==2.1.0
passlib==1.7.4
httpx==0.18.2
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.3.1